… .memento/acronyms.json
… FILESYSTEM_REFACTORING_PROJECT_SUMMARY.md
… package.json
… src/commands/__tests__/acronym.test.ts
… src/commands/__tests__/add.test.ts
… src/commands/__tests__/init-global.test.ts
… src/commands/__tests__/init.test.ts
… src/lib/__tests__/MementoScope.test.ts
… src/lib/__tests__/StarterPackManager.test.ts
… src/lib/__tests__/commandGenerator.test.ts
… src/lib/__tests__/componentInstaller.test.ts
… src/lib/__tests__/configManager.test.ts
… src/lib/__tests__/contracts.test.ts
… src/lib/__tests__/directoryManager.test.ts
… src/lib/__tests__/metadata.test.ts
… src/lib/__tests__/projectDetector.test.ts
… src/lib/__tests__/ticketManager.test.ts
… src/lib/__tests__/updateManager.test.ts
… src/lib/adapters/MemoryFileSystemAdapter.ts
… src/lib/adapters/NodeFileSystemAdapter.ts
… src/lib/adapters/README.md
… src/lib/adapters/__tests__/FileSystemAdapter.test.ts
… src/lib/adapters/index.ts
… src/lib/commandGenerator.ts
… src/lib/componentInstaller.ts
… src/lib/configManager.ts
… src/lib/directoryManager.ts
… src/lib/hooks/HookConfigLoader.ts
… src/lib/hooks/HookFileManager.ts
… src/lib/hooks/HookManager.ts